WIP: html2md: convert newlines to a single space inside inlines
The text content inside an HTML heading element may span multiple lines,
but a Markdown heading can only be written as a single line.

TODO: This adds a new nlToSpace field to printer for now, but maybe
this can be done in a simpler way. See how the rest goes.